fix: change filename to avoid a conflict when api-extractor run
