+++ layout/index.ts
@@ -1,5 +1,5 @@
 import { SmartBannerView } from './view/smart-banner-view'
-import { SmartBannerViewData } from './view/types'
+import { SmartBannerViewData } from './view/data-types'
 
 export {
   SmartBannerView,
