+++ layout/view/action-button/index.ts
@@ -1,4 +1,4 @@
-import { SmartBannerViewData } from '../types';
+import { SmartBannerViewData } from '../data-types';
 
 import styles from './styles.module.scss';
