+++ layout/view/smart-banner-view.ts
@@ -1,5 +1,5 @@
 import { Position } from '@adjustcom/smart-banner-sdk/src/data/types';
-import { SmartBannerViewData } from './types';
+import { SmartBannerViewData } from './data-types';
 import { BannerBody } from './banner-body';
 
 import styles from './styles.module.scss';
