+++ sdk/src/data/converters/smart-banner-for-view.ts
@@ -1,5 +1,5 @@
 import { SmartBannerData } from '../types';
-import { SmartBannerViewData } from '../../../../layout/view/types';
+import { SmartBannerViewData } from '@adjustcom/smart-banner-sdk-layout/view/data-types';
 
 /**
  * Converts SmartBannerData to SmartBannerViewData
